+++ mono_cli/lib/src/system_io/group_store_fs.dart
@@ -1,0 +1,24 @@
+import 'package:mono_cli/mono_cli.dart';
+
+class FileGroupStore implements GroupStore {
+  FileGroupStore(this._folder);
+
+  final ListConfigFolder _folder;
+
+  @override
+  Future<void> deleteGroup(String groupName) => _folder.delete(groupName);
+
+  @override
+  Future<bool> exists(String groupName) => _folder.exists(groupName);
+
+  @override
+  Future<List<String>> listGroups() => _folder.listNames();
+
+  @override
+  Future<List<String>> readGroup(String groupName) =>
+      _folder.readList(groupName);
+
+  @override
+  Future<void> writeGroup(String groupName, List<String> members) =>
+      _folder.writeList(groupName, members);
+}